+++ tests/tests.sh
@@ -8,3 +8,7 @@ nvm use v0.4.0
 node ./index.js
 nvm use v0.4.3
 node ./index.js
+nvm use v0.4.7
+node ./index.js
+nvm use v0.4.8
+node ./index.js