+++ History.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next release
+
+LastFmUpdate can accept any parameter to have passed through to Last.Fm. This allows artist/track info to be supplied without having to use the track object from RecentTrackStream.
+
 ## 0.6.2
 
 Fixed path issue (regression)