+++ History.md
@@ -2,8 +2,9 @@
 
 ## Next release
 
+* Fixed bug where success/error handlers on lastfm.info and lastfm.update were being fired twice.
 * Configurable user-agent string.
-* LastFmUpdate can accept any parameter to have passed through to Last.Fm. This allows artist/track info to be supplied without having to use the track object from RecentTrackStream.
+* LastFmUpdate can now accept any parameter. This will be passed through to Last.Fm. This allows artist/track info to be supplied without having to use the track object from RecentTrackStream.
 
 ## 0.6.2