+++ History.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.3
+
+* Fixed bug where success/error handlers on lastfm.info and lastfm.update were being fired twice.
+* Configurable user-agent string.
+* LastFmUpdate can now accept any parameter. This will be passed through to Last.Fm. This allows artist/track info to be supplied without having to use the track object from RecentTrackStream.
+
 ## 0.6.2
 
 Fixed path issue (regression)